+++ library/dash/src/test/java/com/google/android/exoplayer2/source/dash/manifest/DashManifestParserTest.java
@@ -48,6 +48,7 @@ public class DashManifestParserTest {
       "media/mpd/sample_mpd_unknown_mime_type";
   private static final String SAMPLE_MPD_SEGMENT_TEMPLATE = "media/mpd/sample_mpd_segment_template";
   private static final String SAMPLE_MPD_EVENT_STREAM = "media/mpd/sample_mpd_event_stream";
+  private static final String SAMPLE_MPD_IMAGES = "media/mpd/sample_mpd_images";
   private static final String SAMPLE_MPD_LABELS = "media/mpd/sample_mpd_labels";
   private static final String SAMPLE_MPD_ASSET_IDENTIFIER = "media/mpd/sample_mpd_asset_identifier";
   private static final String SAMPLE_MPD_TEXT = "media/mpd/sample_mpd_text";
@@ -190,6 +191,23 @@ public class DashManifestParserTest {
         new ProgramInformation(
             "MediaTitle", "MediaSource", "MediaCopyright", "www.example.com", "enUs");
     assertThat(manifest.programInformation).isEqualTo(expectedProgramInformation);
+  }
+
+  @Test
+  public void parseMediaPresentationDescription_images() throws IOException {
+    DashManifestParser parser = new DashManifestParser();
+    DashManifest manifest =
+        parser.parse(
+            Uri.parse("https://example.com/test.mpd"),
+            TestUtil.getInputStream(
+                ApplicationProvider.getApplicationContext(), SAMPLE_MPD_IMAGES));
+
+    AdaptationSet adaptationSet = manifest.getPeriod(0).adaptationSets.get(0);
+    Format format = adaptationSet.representations.get(0).format;
+
+    assertThat(format.sampleMimeType).isEqualTo("image/jpeg");
+    assertThat(format.width).isEqualTo(320);
+    assertThat(format.height).isEqualTo(180);
   }
 
   @Test